encounters post layout & nav bar
Updated the post format for “encounters”—now includes image specified
in YAML “image” field and links to the main encounters article,
specified in the YAML “artlink” field. See 2014-04-30-testencounter.md
for format.

Note: even though artlink is pointing to a .md file, it has to be
specified in YAML as .html

Also updated the dropdown nav to reflect site structure.